+++ src/app/pages/service/dtos/connected-devices.dto.ts
@@ -1,0 +1,17 @@
+export interface WayosGetDeviceOnlineUser {
+    auth_type: number;
+    conn_type: number;
+    down_flow: number;
+    ip: string;
+    login_at: number;
+    logout_at: number;
+    mac: string;
+    mfr: number;
+    name: string;
+    os: number;
+    sn: string;
+    ssid: string;
+    status: number;
+    ua: string;
+    up_flow: number;
+}